+++ src/LaravelTypedI18n.php
@@ -1,5 +1,0 @@
-<?php
-
-namespace YuriiDiachuk\LaravelTypedI18n;
-
-class LaravelTypedI18n {}